requirements.txt
absl-py==0.11.0
alembic==1.4.1
altair==4.1.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
azure-core==1.9.0
azure-storage-blob==12.6.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.1
black==20.8b1
bleach==3.2.1
blinker==1.4
bravado==11.0.2
bravado-core==5.17.0
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.7.0
cryptography==2.9.2
cycler==0.10.0
databricks-cli==0.14.1
decorator==4.4.2
defusedxml==0.6.0
docker==4.4.0
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
fsspec==0.8.5
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
glob2==0.7
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
gunicorn==20.0.4
gym==0.17.3
hydra-core==1.0.4
idna==2.9
imageio==2.9.0
importlib-metadata==3.3.0
importlib-resources==3.3.0
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jsonref==0.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.3.1
libarchive-c==2.9
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mlflow==1.12.1
monotonic==1.5
more-itertools==8.6.0
msgpack==1.0.1
msrest==0.6.19
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
neptune-client==0.4.129
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
omegaconf==2.0.5
opencv-python==4.4.0.46
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.0
pathspec==0.8.1
pexpect==4.8.0
pfrl==0.2.1
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
pluggy==0.13.1
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
prompt-toolkit==3.0.5
protobuf==3.14.0
psutil==5.7.0
ptyprocess==0.6.0
py==1.10.0
py3nvml==0.2.6
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pydeck==0.5.0
pyflakes==2.2.0
pyglet==1.5.0
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.1
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-lightning==1.0.4
pytorch-lightning-bolts==0.2.5
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
querystring-parser==1.2.4
regex==2020.11.13
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel-yaml==0.15.87
scikit-image==0.18.0
scikit-learn==0.23.2
scikit-video==1.1.11
scipy==1.5.4
Send2Trash==1.5.0
simplejson==3.17.2
six==1.14.0
smmap==3.0.4
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
streamlit==0.73.0
swagger-spec-validator==2.7.3
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.12.8
toml==0.10.2
toolz==0.11.1
torch==1.6.0
torch-summary==1.4.3
torchvision==0.7.0
tornado==6.1
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.8
validators==0.18.2
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xmltodict==0.12.0
zipp==3.4.0
